backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
redis
razorpay
google-generativeai
scikit-learn
pandas